New companion script to kl_distance.pl Will take the KL distance between two files (A & B) and then try to see how significant that is. First it splits file A into two new files, each containing half of the sequences (randomly chosen) from the parent fie. For up to N times, it calculates the KL distance between these two split files and counts how many times the original distance was exceeded.
If you set N (the 4th parameter on the command line) to be 100, then you can simply say
that if the original distance was exceeded more than 5 times out of 100 then the original
KL distance was not significant.

It then repeats this shuffling for the 2nd file (which can give different results if there
is more or less heterogeneity in this file). Finally it randomizes the DNA of each
file (preserving mononucleotide frequencies) and asks how many times (out of N randomizations)
the original KL distance is exceeded.

Can be slow with large files and/or large word sizes.